pyproject.toml
[project]
name = "youtube-downloader"
version = "0.1.0"
description = ""
authors = [
    {name = "XetPy1030",email = "xetpy1030@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram (>=3.20.0.post0,<4.0.0)",
    "tortoise-orm[asyncpg] (>=0.25.1,<0.26.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "aerich (>=0.9.1,<0.10.0)",
    "yt-dlp (>=2024.1.0,<2025.0.0)",
    "loguru (>=0.7.2,<0.8.0)",
    "pydantic (>=2.5.0,<3.0.0)",
    "pydantic-settings (>=2.1.0,<3.0.0)",
    "httpx (>=0.26.0,<0.27.0)",
    "pillow (>=10.1.0,<11.0.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.aerich]
tortoise_orm = "db_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
